Fix remaining size_t vs int mismatch in AVC
git-svn-id: https://svn.osgeo.org/gdal/trunk@31625 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/avc/avc_e00parse.c
@@ -1273,7 +1273,7 @@ char  **AVCE00ParseNextPrjLine(AVCE00ParseInfo *psInfo, const char *pszLine)
         int  iLastLine, nNewLen;
 
         iLastLine = CSLCount(psInfo->cur.papszPrj) - 1;
-        nNewLen = strlen(psInfo->cur.papszPrj[iLastLine])+strlen(pszLine)-1+1;
+        nNewLen = (int)strlen(psInfo->cur.papszPrj[iLastLine])+strlen(pszLine)-1+1;
         if (iLastLine >= 0)
         {
             psInfo->cur.papszPrj[iLastLine] =
